submissions/hjlees/src/pyproject.toml
[project]
name = "agora-todo-assistant"
version = "1.0.0"
description = "Voice-based todo assistant using Agora Conversational AI, Appwrite, Groq, ElevenLabs, and HeyGen"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "flask>=3.0.0",
    "python-dotenv>=1.0.0",
    "appwrite>=5.0.0",
    "groq>=0.11.0",
    "elevenlabs>=1.11.0",
    "agora-python-server-sdk>=2.0.0",
    "agora-token-builder>=1.0.0",
    "requests>=2.32.0",
    "pydub>=0.25.1",
    "soundfile>=0.12.1",
    "numpy>=1.24.0",
    "pydantic>=2.0.0",
    "python-dateutil>=2.8.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
]

